+++ src/server/devices/shelly/index.ts
@@ -1,0 +1,2 @@
+export { ShellyDevice } from './shellyDevice';
+export { ShellyTrv } from './shellyTrv';