+++ CHANGELOG.md
@@ -5,6 +5,12 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Add Shelly device handling
+* (theimo1221) Add first Shelly Device (Shelly TRV)
+
 ## 2.15.10 (2023-10-08)
 
 * (theimo1221) Add option to not send commands to unavailable Zigbee Devices